Some semi significant commit here. Mostly focused on updating the vignette so Ivans predict function has more substance to it as its current stage, fairly sure its accurate. Used the htmltools r library to get the bettr logo in the vignette too, looks the same as the last code but this is cleaner and easier to write/understand.
I have now upaded the BETtR-package so that the .onattach now has the read.dcf function so when the versions are updated in the description it will display the CURRENT version of the package. Also wrapped it in an 'if(interactive())' so it is no longer "technically wrong" as Keefe said.

Decided to update the README.md too. Pretty self explanatory whats done here. Kept it short and simple for the current version.

And just tacked on 'htmltools' under Suggest due to its use in the vignette.